topdown: Move all the jwt tests into one place
The tests were split up between the gigantic `topdown_test.go` file
and the `tokens_test.go`file. This just moves the unit tests for
stuff in `tokens.go` into `tokens_test.go`.
